Move state gen to state
