+++ beacon-chain/state/stategen/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "replay.go",
         "service.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/stategen",
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/state/stategen",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
         "//beacon-chain/core/state:go_default_library",
