+++ beacon-chain/state/stategen/service.go
